requirements.txt
pycryptodome==3.17
libnum==1.7.1
Cython==0.29.35
numpy==1.21.6

matplotlib==3.5.3
Pympler==1.0.1

pytest==7.3.1

# It is needed to download and install this by hand :(
# https://github.com/JHUISI/charm